+++ CHANGELOG.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 v1.1.4
 ------
-- Add ParserDefinition.callable() decorator to replace env parameter in get_parser
+- Add callable() decorator to replace env parameter in get_parser
 - bool added to primitive types
 
 v1.1.3